添加了默认的排序方式

+++ src/modules/pm-sub/pm-sub.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
+import { BusinessException } from "~/common/exceptions/biz.exception";
 import { PMSubEntity } from "~/entities/pm_sub";
 import { PMSubNetWorkDeviceEntity } from "~/entities/pm_sub_network_device";
 import { paginate } from "~/helper/paginate";
@@ -17,6 +18,7 @@ export class PmSubService {
     private readonly pm_sub_network_entity: Repository<PMSubNetWorkDeviceEntity>
   ) {}
 
+  // TAG(2025-03-17 00:09:46 谭人杰): 子任务接口
   async search() {
     return await paginate(
       this.pm_sub_entity,
@@ -30,29 +32,22 @@ export class PmSubService {
   }
 
   async complete(mId: number, dto: CompleteDto, uId: number) {
+    let entites = await this.findByName_sub(dto.sub.pm_sub_name)
+    if(entites.length > 0 && dto.id != entites?.[0]?.id) throw new BusinessException("500:The name already exists")
     return await this.pm_sub_entity.manager.transaction(async (manager) => {
-      console.log({
-        id: dto.id,
-        ...dto.sub,
-        ...(dto.id ? { updateBy: uId }: { createBy: uId }),
-        pm_master: {
-          id: mId,
-        },
-      },);
-      
       let sub_entity = await manager.upsert(
         PMSubEntity,
         {
           id: dto.id,
           ...dto.sub,
-          ...(dto.id ? { updateBy: 1 }: { createBy: 1 }),
+          ...(dto.id ? { updateBy: 1 } : { createBy: 1 }),
           pm_master: {
             id: mId,
           },
         },
         ["id"]
       );
-      let sub_id = sub_entity.identifiers?.[0]?.id
+      let sub_id = sub_entity.identifiers?.[0]?.id;
       await manager
         .createQueryBuilder()
         .delete()
@@ -68,13 +63,13 @@ export class PmSubService {
           id: sub_id,
         },
         createBy: 1,
-        updateBy: 1
+        updateBy: 1,
       }));
       await manager.save(PMSubNetWorkDeviceEntity, networks);
       return {
         id: sub_id,
         ...dto.sub,
-        ...(dto.id ? { updateBy: dto.id }: { createBy: dto.id }),
+        ...(dto.id ? { updateBy: dto.id } : { createBy: dto.id }),
         pm_master: {
           id: mId,
         },
@@ -114,9 +109,16 @@ export class PmSubService {
   }
 
   async delete_sub(id: number) {
-    return await this.pm_sub_entity.delete(id)
+    return await this.pm_sub_entity.delete(id);
   }
 
+  async findByName_sub(name: string) {
+    return await this.pm_sub_entity.findBy({
+      pm_sub_name: name
+    });
+  }
+
+  // TAG(2025-03-17 00:10:11 谭人杰):子任务-设备接口
   async create_network_device(
     sId: number,
     dto: SubNetWorkDeviceDto,
@@ -155,6 +157,6 @@ export class PmSubService {
   }
 
   async delete_network_device(id: number) {
-    return await this.pm_sub_network_entity.delete(id)
+    return await this.pm_sub_network_entity.delete(id);
   }
 }
